wired up search box on music library to API, relocated API to be in the libraryController so the code is set up for pieces to split off and be modular
