+++ server/routes/api/library.js
@@ -3,4 +3,6 @@ const libraryController = require('../../controllers/libraryController');
 
 router.route('/').get(libraryController.findAll);
 
+router.route('/search').get(libraryController.searchLibrary);
+
 module.exports = router;
